Fix correctness defects in the vendored e3sm_diags_parser
Prepare the parser to stand in for the create_diagnostic regexes. It was
not usable against EAMxx field names before this. No EAMxx code calls it
yet, so nothing outside the package changes behavior.

Crashes and hangs:
- parse_prefix_expression captured the operator but never advanced past
  it, so parse_expression re-dispatched the same token and recursed
  until the stack overflowed. "-T" segfaulted.
- read_to_delim looped until it found the closing quote with no
  end-of-input check, so an unterminated string literal hung forever.
- Illegal tokens did not advance the lexer, so iterating next_token to
  EndofFile never terminated.
- "1e" ran position_ past the end of the input and threw out_of_range
  from substr.

Silently wrong results, which matter most here because a typo in a
user's YAML would have quietly computed a different diagnostic:
- Unknown tokens were printed to std::cout and execution continued, and
  parse() never required end-of-input. "bc_a1" parsed as "bc_a", "qc !=
  0" as "qc", "T @ x" as "t". Both now record errors and throw
  ParserError.
- A number was only classified Float if it contained 'e', so "0.5"
  became an Integer token and stoi truncated it to 0. "where(qc > 0.5)"
  silently became "where(qc > 0)".
- stoi/stod results were not checked against the full literal.

EAMxx field names:
- The lexer lower-cased its whole input, so T_mid, LiqWaterPath and
  AeroComCldTop named nothing real. Keywords are now lower-case only,
  matching Python.
- Identifiers excluded digits, shattering bc_a1, so4_a2, num_a3 and O3
  into several tokens. Now Python rules: no leading digit, digits
  allowed thereafter.
- '!' was never lexed, so NotEqual could not be produced despite being
  wired through precedence and printing.

Also: parenthesized grouping never worked, because
parse_grouped_expression was defined but never registered as a prefix
parse function, so operator precedence could not be overridden at all;
binary_op_to_string was missing and, or, >= and <=, so to_string threw
on valid ASTs; FloatLiteral is now double and prints the shortest
round-tripping form rather than "%e" from a float.

Tests go from 3 cases to 18 (331 assertions), one per defect plus
realistic expressions. Warning-clean under -Wall -Wextra -Wpedantic and
clean under -fsanitize=address,undefined. README documents each
divergence from upstream with its repro; all are candidates for
upstreaming.

+++ components/eamxx/src/share/io/edp/include/edp/ast.hpp
@@ -59,7 +59,9 @@ struct StringLiteral {
 };
 
 struct FloatLiteral {
-  float value;
+  // NOTE: upstream stored a `float`. Diagnostic thresholds and pressure levels
+  //       need double precision, so the value is kept as parsed.
+  double value;
 };
 
 struct IntegerLiteral {
